feat: add RoomList asyncboundary
