+++ packages/client/src/components/base/ErrorFallback.tsx
@@ -1,5 +1,6 @@
 import { useEffect } from 'react';
 import { useQueryClient } from '@tanstack/react-query';
+import styled from '@emotion/styled';
 
 interface Props {
   message: string;
@@ -14,11 +15,33 @@ const ErrorFallback = ({ message, queryKey }: Props) => {
   }, [queryClient, queryKey]);
 
   return (
-    <div>
-      <div>ERROR</div>
-      <div>{message}</div>
-    </div>
+    <Container>
+      <Text>
+        <h1>ERROR</h1>
+        <h2>{message}</h2>
+      </Text>
+    </Container>
   );
 };
+
+const Container = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+  width: 100%;
+`;
+
+const Text = styled.div`
+  color: red;
+  h1 {
+    font-weight: 700;
+    font-size: 4rem;
+  }
+  h2 {
+    font-weight: 700;
+    font-size: 2rem;
+  }
+`;
 
 export default ErrorFallback;
