+++ packages/client/src/components/common/Loader/Loader.tsx
@@ -1,4 +1,4 @@
-import { NormalColorType } from '~/styles/palette';
+import { NormalColorType } from '~/styles';
 import * as S from './Loader.styles';
 
 export interface LoaderProps {
